Cleanup + refactoring delete image to untag image
