+++ commands/azureCommands/acr-build.ts
@@ -1,22 +1,14 @@
-import { ContainerRegistryManagementClient } from 'azure-arm-containerregistry/lib/containerRegistryManagementClient';
 import { DockerBuildRequest } from "azure-arm-containerregistry/lib/models/dockerBuildRequest";
 import { Registry } from 'azure-arm-containerregistry/lib/models/registry';
-import { BlobService, createBlobServiceWithSas } from "azure-storage";
-import * as fs from 'fs';
-import * as os from 'os';
-import * as process from 'process';
-import * as tar from 'tar';
-import * as url from 'url';
 import * as vscode from "vscode";
 import { IAzureQuickPickItem } from 'vscode-azureextensionui';
 import { ext } from '../../extensionVariables';
-import { getBlobInfo, getResourceGroupName, streamLogs } from "../../utils/Azure/acrTools";
+import { getResourceGroupName, streamLogs } from "../../utils/Azure/acrTools";
 import { AzureUtilityManager } from "../../utils/azureUtilityManager";
 import { resolveFileItem } from '../build-image';
+import { getTempSourceArchivePath, uploadSourceCode } from '../utils/OutputChannel';
 import { quickPickACRRegistry, quickPickNewImageName, quickPickSubscription } from '../utils/quick-pick-azure';
 
-const idPrecision = 6;
-const vcsIgnoreList = ['.git', '.gitignore', '.bzr', 'bzrignore', '.hg', '.hgignore', '.svn'];
 const status = vscode.window.createOutputChannel('ACR Build status');
 
 // Prompts user to select a subscription, resource group, then registry from drop down. If there are multiple folders in the workspace, the source folder must also be selected.
@@ -47,9 +39,9 @@ export async function queueBuild(dockerFileUri?: vscode.Uri): Promise<void> {
     }
     const dockerItem = await resolveFileItem(folder, dockerFileUri, "DockerFile");
     const sourceLocation: string = folder.uri.path;
-    const tarFilePath = getTempSourceArchivePath();
+    const tarFilePath = getTempSourceArchivePath(status);
 
-    const uploadedSourceLocation = await uploadSourceCode(client, registry.name, resourceGroupName, sourceLocation, tarFilePath, folder);
+    const uploadedSourceLocation = await uploadSourceCode(status, client, registry.name, resourceGroupName, sourceLocation, tarFilePath, folder);
     status.appendLine("Uploaded Source Code to " + tarFilePath);
 
     const runRequest: DockerBuildRequest = {
@@ -66,51 +58,4 @@ export async function queueBuild(dockerFileUri?: vscode.Uri): Promise<void> {
     status.appendLine("Schedule Run " + run.runId);
 
     streamLogs(registry, run, status, client);
-}
-
-async function uploadSourceCode(client: ContainerRegistryManagementClient, registryName: string, resourceGroupName: string, sourceLocation: string, tarFilePath: string, folder: vscode.WorkspaceFolder): Promise<string> {
-    status.appendLine("   Sending source code to temp file");
-    let source = sourceLocation.substring(1);
-    let current = process.cwd();
-    process.chdir(source);
-    fs.readdir(source, (err, items) => {
-        items = filter(items);
-        tar.c(
-            {},
-            items
-        ).pipe(fs.createWriteStream(tarFilePath));
-        process.chdir(current);
-    });
-
-    status.appendLine("   Getting Build Source Upload Url ");
-    let sourceUploadLocation = await client.registries.getBuildSourceUploadUrl(resourceGroupName, registryName);
-    let upload_url = sourceUploadLocation.uploadUrl;
-    let relative_path = sourceUploadLocation.relativePath;
-
-    status.appendLine("   Getting blob info from Upload Url ");
-    // Right now, accountName and endpointSuffix are unused, but will be used for streaming logs later.
-    let { accountName, endpointSuffix, containerName, blobName, sasToken, host } = getBlobInfo(upload_url);
-    status.appendLine("   Creating Blob Service ");
-    let blob: BlobService = createBlobServiceWithSas(host, sasToken);
-    status.appendLine("   Creating Block Blob ");
-    blob.createBlockBlobFromLocalFile(containerName, blobName, tarFilePath, (): void => { });
-    return relative_path;
-}
-
-function getTempSourceArchivePath(): string {
-    /* tslint:disable-next-line:insecure-random */
-    let id = Math.floor(Math.random() * Math.pow(10, idPrecision));
-    status.appendLine("Setting up temp file with 'sourceArchive" + id + ".tar.gz' ");
-    let tarFilePath = url.resolve(os.tmpdir(), `sourceArchive${id}.tar.gz`);
-    return tarFilePath;
-}
-
-function filter(list: string[]): string[] {
-    let result = [];
-    for (let file of list) {
-        if (vcsIgnoreList.indexOf(file) === -1) {
-            result.push(file);
-        }
-    }
-    return result;
 }
